Huge refactor of how SetupTeardown proxy talks to the nodes before during and after session. This should prevent a lot of thread locking when video start/stop/update happens.

+++ SeleniumGridExtras/src/main/java/com/groupon/seleniumgridextras/grid/proxies/sessions/threads/NodeRestartCallable.java
@@ -2,11 +2,21 @@ package com.groupon.seleniumgridextras.grid.proxies.sessions.threads;
 
 import com.google.gson.JsonObject;
 import com.groupon.seleniumgridextras.grid.proxies.SetupTeardownProxy;
+import com.groupon.seleniumgridextras.tasks.GridStatus;
+import com.groupon.seleniumgridextras.tasks.config.TaskDescriptions;
+import com.groupon.seleniumgridextras.utilities.json.JsonCodec;
+import com.groupon.seleniumgridextras.utilities.json.JsonParserWrapper;
+import com.groupon.seleniumgridextras.utilities.threads.CommonThreadPool;
+import com.groupon.seleniumgridextras.utilities.threads.RemoteGridExtrasAsyncCallable;
 import org.apache.log4j.Logger;
 import org.openqa.grid.internal.TestSession;
 import org.openqa.grid.selenium.proxy.DefaultRemoteProxy;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.Callable;
+import java.util.concurrent.Future;
 
 
 public class NodeRestartCallable implements Callable {
@@ -35,11 +45,11 @@ public class NodeRestartCallable implements Callable {
             logger.error(e);
         }
 
+        Map status = null;
         try {
 
 
-
-            if (this.proxy.isBusy() || this.proxy.getRegistry().getNewSessionRequestCount() != 0){
+            if (this.proxy.isBusy() || this.proxy.getRegistry().getNewSessionRequestCount() != 0) {
                 String message = String.format("Proxy %s is currently %s busy and has there are %s items in the queue. Will not attempt to reboot node until the grid is free",
                         this.proxy.getId(),
                         (this.proxy.isBusy() ? "" : "not"),
@@ -49,7 +59,34 @@ public class NodeRestartCallable implements Callable {
                 return message;
             }
 
-            JsonObject status = proxy.callRemoteGridExtras("grid_status");
+            Future<String> f = CommonThreadPool.startCallable(
+                    new RemoteGridExtrasAsyncCallable(
+                            proxy.getRemoteHost().getHost(),
+                            3000,
+                            TaskDescriptions.Endpoints.GRID_STATUS,
+                            new HashMap<String, String>()));
+
+            String response = "";
+            try {
+                response = f.get();
+                logger.debug(response);
+            } catch (Exception e) {
+                logger.error(
+                        String.format(
+                                "Error getting the %s endpoint for proxy %s ",
+                                TaskDescriptions.Endpoints.GRID_STATUS,
+                                proxy.getId()),
+                        e);
+            }
+
+
+            if (response.equals("")){
+                String error = "Something went wrong when asking for status from " + proxy.getId();
+                logger.error(error);
+                return error;
+            }
+
+            status = JsonParserWrapper.toHashMap(response);
 
             if (status == null) {
                 String message = String.format("Problem communicating with %s, will not attempt to reboot", proxy.getRemoteHost().getHost());
@@ -57,8 +94,17 @@ public class NodeRestartCallable implements Callable {
                 return message;
             }
 
-            int sessionsStarted = status.get("node_sessions_started").getAsInt();
-            int sessionLimit = status.get("node_sessions_limit").getAsInt();
+
+            int recordedSessions;
+            if (status.containsKey(GridStatus.DEPRECATED_STARTED_SESSIONS_KEY)) {
+                logger.warn(GridStatus.DEPRECATION_WARNING + " for node " + this.proxy.getId());
+                recordedSessions = ((Double) status.get(GridStatus.DEPRECATED_STARTED_SESSIONS_KEY)).intValue();
+            } else {
+                recordedSessions = ((List) status.get(JsonCodec.WebDriver.Grid.RECORDED_SESSIONS)).size();
+            }
+
+
+            int sessionLimit = ((Double) status.get(JsonCodec.WebDriver.Grid.NODE_SESSIONS_LIMIT)).intValue();
 
             if (sessionLimit == 0) {
                 String message = String.format("Node %s with proxy %s is set to never reboot, skipping this step",
@@ -67,15 +113,17 @@ public class NodeRestartCallable implements Callable {
                 return message;
             }
 
-            if (sessionsStarted >= sessionLimit) {
+            if (recordedSessions >= sessionLimit) {
                 String message = String.format("Node %s has executed has executed %s sessions, the limit is %s so it is time to reboot it",
-                        proxy.getRemoteHost().getHost(), sessionsStarted, sessionLimit);
+                        proxy.getRemoteHost().getHost(), recordedSessions, sessionLimit);
 
                 logger.info(message);
                 proxy.setAvailable(false);
                 proxy.setRestarting(false);
+
+                String host = proxy.getRemoteHost().getHost(); //Grabbing host before un-register call, just in case it gets set to null
                 proxy.stopGridNode();
-                proxy.rebootGridExtrasNode();
+                SetupTeardownProxy.rebootGridExtrasNode(host);
 
 
                 return message;
@@ -83,8 +131,8 @@ public class NodeRestartCallable implements Callable {
         } catch (Exception e) {
             //Capture any unexpected exception and print it to log. Currently the hub will completely ignore any
             //Exception, this way at least we can try to debug it.
-            String message = String.format("Something didn't go right when trying to reboot node %s : %s",
-                    proxy.getRemoteHost().getHost(), e.getMessage());
+            String message = String.format("Something didn't go right when trying to reboot node %s : %s. \nReply from node:\n%s",
+                    proxy.getRemoteHost().getHost(), e.getMessage(), "" + status);
             logger.error(message, e);
             return message;
         }
